Chore : save winner By Event;

+++ back/src/main/resources/db/migration/V1.0__init.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS `raffles`
     end_time        dateTime        NOT NULL,
     is_end          boolean         NOT NULL    DEFAULT false,
     is_paid         boolean         NOT NULL    DEFAULT false,
-    settlement      bigint          NULL,
+    settlement      double          NULL,
     created_at      dateTime        NOT NULL    DEFAULT now(),
     updated_at      dateTime        NOT NULL    DEFAULT now()
 );
